feat: set fetcher at once in _app
#30
