+++ client/src/pages/r/[sub].tsx
@@ -9,18 +9,11 @@ import useSWR from 'swr';
 const SubPage = () => {
     const [ownSub, setOwnSub] = useState(false);
     const { authenticated, user } = useAuthState();
-    const fetcher = async (url: string) => {
-        try {
-            const res = await axios.get(url);
-            return res.data;
-        } catch (error: any) {
-            throw error.response.data;
-        }
-    };
+    
     const fileInputRef = useRef<HTMLInputElement>(null);
     const router = useRouter();
     const subName = router.query.sub;
-    const { data: sub, error } = useSWR(subName ? `/subs/${subName}` : null, fetcher);
+    const { data: sub, error } = useSWR(subName ? `/subs/${subName}` : null);
     console.log('sub', sub);
 
     useEffect(() => {
